Basic start
I added the text fields which allows the user to enter both roots and their translations. This creates two arrays (one array for the roots, another for the translations).  Using DOM a title (an h3 and p tag within a div)  is generated for each root given, which lists the root and the translation. So far they are not displayed inline Later inlection tables will be added under these headwords.