Vastly cleaned the code. Move
I removed a lot of repetitive blocks of code. Instead of having a new block of code for each new instance of a "copied adjective/noun/verb", this is now handled by a few lines of code in a for loop. I also placed the lengthy arrays of English words into their own files which are then imported into script.js, before I had them all in script.js which made it very untidy